requirements.txt
Django==3.1.7
djangorestframework==3.12.4
drf-yasg==1.20.0